refactor: refactor Filter : make it reuseable. refactor all table: use shared Filter.
