+++ src/components/views/DatabaseView/shared/index.ts
@@ -1,2 +1,3 @@
 // 共用組件的統一導出
-export { default as EditableCell } from "./EditableCell/index";
+export { default as EditableCell } from "./EditableCell";
+export { Filter } from "./Filter";
